refactor: move some logic out of main
